backend/requirements.txt
Django==6.0.4
django_auditlog==3.4.1
django_environ==0.12.0
django_import_export==4.4.0
django_multiselectfield==1.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
weasyprint==68.1
